eyeCU-FE-master/newTestServer/requirements.txt
alembic==1.7.6
aniso8601==9.0.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
cors==1.0.1
dlib==19.23.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.6.0
Flask==1.1.4
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Migrate==2.1.1
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.4.0
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.22.2
Pillow==9.0.1
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.1
PySocks==1.7.1
pytz==2021.3
requests==2.27.1
requests-file==1.5.1
six==1.16.0
SQLAlchemy==1.4.31
tldextract==3.2.0
urllib3==1.26.8
Werkzeug==1.0.1
WTForms==3.0.1
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
